refactor imports so pwg conditionally uses wasm or ffis depending on the flag
