+++ src/acvm_interop/pwg/gadget_call.rs
@@ -1,5 +1,5 @@
 use super::merkle::MerkleTree;
-use crate::barretenberg_rs::Barretenberg;
+use crate::Barretenberg;
 use acvm::acir::{circuit::gate::GadgetCall, native_types::Witness, OPCODE};
 use acvm::pwg::{self, input_to_value};
 use acvm::FieldElement;
